Extract DirectWrite font picker logic into separate class
