+++ foo_ui_columns/foo_ui_columns.vcxproj
@@ -269,6 +269,7 @@
     <ClCompile Include="file_info_utils.cpp" />
     <ClCompile Include="font_manager.cpp" />
     <ClCompile Include="font_manager_v3.cpp" />
+    <ClCompile Include="font_picker.cpp" />
     <ClCompile Include="gdi.cpp" />
     <ClCompile Include="icons.cpp" />
     <ClCompile Include="item_details_format_parser.cpp" />
@@ -448,6 +449,7 @@
     <ClInclude Include="file_info_utils.h" />
     <ClInclude Include="font_manager_v3.h" />
     <ClInclude Include="font_manager_data.h" />
+    <ClInclude Include="font_picker.h" />
     <ClInclude Include="gdi.h" />
     <ClInclude Include="gdiplus.h" />
     <ClInclude Include="get_msg_hook.h" />
