+++ foo_ui_columns/foo_ui_columns.vcxproj.filters
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <Filter Include="Resources">
@@ -626,6 +626,9 @@
     <ClCompile Include="item_details_format_parser.cpp">
       <Filter>Item details</Filter>
     </ClCompile>
+    <ClCompile Include="font_picker.cpp">
+      <Filter>Utilities</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h">
@@ -924,6 +927,9 @@
       <Filter>Item details</Filter>
     </ClInclude>
     <ClInclude Include="lexy_user_config.hpp" />
+    <ClInclude Include="font_picker.h">
+      <Filter>Utilities</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include=".clang-format" />
